# openapi.yaml
openapi: 3.0.0
info:
  title: User Authentication and Account API
  version: 1.0.0
  description: API for user sign-up, sign-in, email verification, password management, and session handling.
servers:
  - url: /api # Assuming your API routes are under /api
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        details: # Optional, for more specific error info
          type: string
          nullable: true
...
    FeatureResponse: # Renamed from Feature and added id
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the feature.
        name:
          type: string
          description: Name of the feature.
        included:
          type: boolean
          description: Whether the feature is included.
      required:
        - id
        - name
        - included
    FeatureInput:
      type: object
      properties:
        id:
          type: string
          nullable: true
          description: Unique identifier for the feature (used for updates, optional for new features within a package update).
        name:
          type: string
          description: Name of the feature.
        included:
          type: boolean
          description: Whether the feature is included.
      required:
        - name
        - included
    CategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the category.
        name:
          type: string
          description: Name of the category.
        icon:
          type: string
          nullable: true
          description: Icon for the category (e.g., SVG content or a class name).
      required:
        - id
        - name
    SubcategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the subcategory.
        name:
          type: string
          description: Name of the subcategory.
        categoryId:
          type: string
          description: ID of the parent category.
      required:
        - id
        - name
        - categoryId
...
    PackageBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          nullable: true
        bgColor:
          type: string
          nullable: true
      required:
        - id
        - name
        - price
        - categoryId
        - subcategoryId
...
    Subcategory: # Updated to be SubcategoryFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        categoryId:
          type: string
        category:
          $ref: '#/components/schemas/CategoryBrief'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/PackageBrief'
      required:
        - id
        - name
        - categoryId
    CreateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the subcategory.
        categoryId:
          type: string
          description: ID of the parent category.
      required:
        - name
        - categoryId
    UpdateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
          description: New name for the subcategory.
        categoryId:
          type: string
          description: New parent category ID.
...
    Package: # Updated to be PackageFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          nullable: true
        bgColor:
          type: string
          nullable: true
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureResponse'
        category:
          $ref: '#/components/schemas/CategoryBrief'
        subcategory:
          $ref: '#/components/schemas/SubcategoryBrief'
      required:
        - id
        - name
        - price
        - categoryId
        - subcategoryId
    CreatePackageRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          default: false
        bgColor:
          type: string
          nullable: true
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureInput' # id should not be provided for new features
          description: List of features for the package.
      required:
        - name
        - price
        - categoryId
        - subcategoryId
        - features
...
  securitySchemes:
    cookieAuth: # For NextAuth.js session cookies
      type: apiKey
      in: cookie
      name: next-auth.session-token # Or __Secure-next-auth.session-token for HTTPS
security:
  - cookieAuth: [] # Indicates that most authenticated routes use cookieAuth
paths:
...
  /product/subcategories:
    post:
      summary: Create a new subcategory
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubcategoryRequest'
      responses:
        '201':
          description: Subcategory created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory' # Returns full subcategory
        '400':
          description: Invalid input (e.g., name missing, parent category not found, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parent category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., subcategory with this name already exists in the parent category).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all subcategories
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - name: categoryId
          in: query
          required: false
          description: Filter subcategories by parent category ID.
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved subcategories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subcategory' # Returns full subcategories
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /product/subcategories/{subcategoryId}:
    parameters:
      - name: subcategoryId
        in: path
        required: true
        description: ID of the subcategory.
        schema:
          type: string
    get:
      summary: Get a specific subcategory by ID
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved the subcategory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory' # Returns full subcategory
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subcategory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a subcategory
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubcategoryRequest'
      responses:
        '200':
          description: Subcategory updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory' # Returns updated full subcategory
        '400':
          description: Invalid input (e.g., parent category not found, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subcategory or target parent category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., subcategory with this name already exists in the target category).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a subcategory
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Subcategory deleted successfully.
        '400':
          description: Subcategory cannot be deleted (e.g., has associated packages).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Subcategory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /product/packages:
    post:
      summary: Create a new package
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePackageRequest'
      responses:
        '201':
          description: Package created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package' # Returns full package with features
        '400':
          description: Invalid input (e.g., missing fields, category/subcategory mismatch, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category or Subcategory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., package with this name already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all packages
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - name: categoryId
          in: query
          required: false
          description: Filter packages by category ID.
          schema:
            type: string
        - name: subcategoryId
          in: query
          required: false
          description: Filter packages by subcategory ID.
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package' # Returns full packages with features
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...